+++ shared/src/chess/rules/diagonal-movement.rule.ts
@@ -1,4 +1,5 @@
-import { Direction } from "./rules";
+import { Coordinate } from "../types";
+import { Direction } from "./movement-rule";
 import {
   StraightMovementRule,
   StraightMovementRuleConfig,
@@ -11,8 +12,9 @@ export class DiagonalMovementRule extends StraightMovementRule {
     collision,
     distance,
     directions,
+    speed = 1,
   }: StraightMovementRuleConfig) {
-    super(moveToEmpty, moveToKill, collision, distance, directions);
+    super(moveToEmpty, moveToKill, collision, distance, directions, speed);
   }
 
   protected possibleDirrections = [
@@ -26,7 +28,7 @@ export class DiagonalMovementRule extends StraightMovementRule {
     y: number,
     diff: number,
     dirrection: Direction
-  ): [number, number] => {
+  ): Coordinate => {
     if (dirrection == Direction.UpLeft) {
       return [x - diff, y - diff];
     }