+++ shared/src/chess/rules/index.ts
@@ -1,7 +1,8 @@
 export * from "./rules";
-
+export * from "./movement-rule";
 export * from "./straight-movement.rule";
 export * from "./diagonal-movement.rule";
 export * from "./horizontal-movement.rule";
 export * from "./vertical-movement.rule";
 export * from "./knight-movement.rule";
+export * from "./position-specific-movement.rule";