+++ shared/src/chess/types.ts
@@ -1,0 +1,1 @@
+export type Coordinate = [number, number]; // [x, y],  when you access board use squares[y, x]